requirements.txt
streamlit
openai
pandas
matplotlib
python-docx
docx2pdf